requirements.txt
toml>=0.10.2
httpx>=0.20.0
Pillow>=8.4.0
aiofiles>=0.7.0
quart>=0.14.1
requests>=2.26.0